Bug-fix in Empirical. Slight change in behaviour of PerHostTransmission; shouldn't have much effect.
git-svn-id: https://openmalaria.googlecode.com/svn/trunk@560 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ model/Surveys.cpp
@@ -33,7 +33,7 @@ void SurveysType::init ()
   const scnXml::Surveys::SurveyTimeSequence& survs = getMonitoring().getSurveys().getSurveyTime();
 
   _surveysTimeIntervals.resize (survs.size() + 1);
-  for (int i = 0; i < survs.size(); i++) {
+  for (size_t i = 0; i < survs.size(); i++) {
     _surveysTimeIntervals[i] = survs[i];
   }
   _surveysTimeIntervals[survs.size()] = -1;
